Handle LoRA loading from queue import
When loaded from a queue import the selected_loras and lora_values are equal in length.
When loaded from the Generate interface the lora_values are the same length as the lora_loaded_names and must be reduced.

This change now makes the assumption that when the two lists are the same length that they are in the correct order.